tech: refactoring SSC integration related stuff
