+++ integrations-dto/src/main/java/com/hp/octane/integrations/dto/general/CIServerInfo.java
@@ -23,73 +23,61 @@ import com.hp.octane.integrations.dto.DTOBase;
 
 public interface CIServerInfo extends DTOBase {
 
-    String getType();
+	String getType();
 
-    /**
-     * Set CI server's type
-     * The type is an unbound string, but please refer to an existing types in CIServerTypes enumeration for a matching existing one, if possible
-     */
-    CIServerInfo setType(String type);
+	/**
+	 * Set CI server's type
+	 * The type is an unbound string, but please refer to an existing types in CIServerTypes enumeration for a matching existing one, if possible
+	 */
+	CIServerInfo setType(String type);
 
-    String getVersion();
+	String getVersion();
 
-    /**
-     * Set CI server's version
-     */
-    CIServerInfo setVersion(String version);
+	/**
+	 * Set CI server's version
+	 */
+	CIServerInfo setVersion(String version);
 
-    String getUrl();
+	String getUrl();
 
-    /**
-     * Set CI server's self location URL
-     */
-    CIServerInfo setUrl(String url);
+	/**
+	 * Set CI server's self location URL
+	 */
+	CIServerInfo setUrl(String url);
 
-    String getInstanceId();
+	String getInstanceId();
 
-    /**
-     * Set unique id of the CI server instance UUID
-     */
-    CIServerInfo setInstanceId(String instanceId);
+	/**
+	 * Set unique id of the CI server instance UUID
+	 */
+	CIServerInfo setInstanceId(String instanceId);
 
-    Long getInstanceIdFrom();
+	Long getInstanceIdFrom();
 
-    /**
-     * Set CI server's instance UUID creation time
-     */
-    CIServerInfo setInstanceIdFrom(Long instanceIdFrom);
+	/**
+	 * Set CI server's instance UUID creation time
+	 */
+	CIServerInfo setInstanceIdFrom(Long instanceIdFrom);
 
-    Long getSendingTime();
+	Long getSendingTime();
 
-    /**
-     * Set CI server's current time
-     */
-    CIServerInfo setSendingTime(Long sendingTime);
+	/**
+	 * Set CI server's current time
+	 */
+	CIServerInfo setSendingTime(Long sendingTime);
 
-    String getImpersonatedUser();
+	String getImpersonatedUser();
 
-    /**
-     * Set CI server's user, that has been given to Octane's CI Plugin to run on behalf of
-     */
-    CIServerInfo setImpersonatedUser(String impersonatedUser);
+	/**
+	 * Set CI server's user, that has been given to Octane's CI Plugin to run on behalf of
+	 */
+	CIServerInfo setImpersonatedUser(String impersonatedUser);
 
-    boolean isSuspended();
+	boolean isSuspended();
 
-    /**
-     * Set CI server's suspension status
-     */
-    CIServerInfo setSuspended(boolean suspended);
-
-    String getSSCURL();
-
-    CIServerInfo setSSCURL(String sscUrl);
-
-    String getSSCBaseAuthToken();
-
-    CIServerInfo setSSCBaseAuthToken(String sscBaseAuthToken);
-
-    long getMaxPollingTimeoutHours();
-
-    CIServerInfo setMaxPollingTimeoutHours(long maxPollingTimeoutHours);
+	/**
+	 * Set CI server's suspension status
+	 */
+	CIServerInfo setSuspended(boolean suspended);
 }
